+++ inst/include/dplyr/comparisons.h
@@ -18,10 +18,6 @@ namespace dplyr {
             return lhs > rhs ;    
         }
         
-        inline bool is_equal(STORAGE lhs, STORAGE rhs ) const {
-            return lhs == rhs ;    
-        }
-        
         inline bool equal_or_both_na( STORAGE lhs, STORAGE rhs ) const {
             return lhs == rhs ;    
         }
@@ -45,10 +41,6 @@ namespace dplyr {
             if (is_na(lhs)) return false;
             if (is_na(rhs)) return true;
             return strcmp( CHAR(lhs), CHAR(rhs) ) > 0;    
-        }
-        
-        inline bool is_equal(SEXP lhs, SEXP rhs ) const {
-            return lhs == rhs ;    
         }
         
         inline bool equal_or_both_na( SEXP lhs, SEXP rhs ) const {
@@ -89,10 +81,6 @@ namespace dplyr {
             
         }
         
-        inline bool is_equal(double lhs, double rhs ) const {
-            return lhs == rhs ;    
-        }
-        
         inline bool equal_or_both_na( double lhs, double rhs ) const {
             return lhs == rhs ||
                 ( is_nan(lhs) && is_nan(rhs) ) ||
@@ -106,16 +94,6 @@ namespace dplyr {
         inline bool is_nan(double x) const {
             return Rcpp::traits::is_nan<REALSXP>(x) ;    
         }
-        
-        // inline void print(double x) const {
-        //     if( is_na(x) ) {
-        //         std::cout << "NA" ;
-        //     } else if( is_nan(x) ) {
-        //         std::cout << "NaN" ;
-        //     } else {
-        //         std::cout << x ;    
-        //     }
-        // }
         
     } ;