+++ tests/testthat/test-group-by.r
@@ -152,3 +152,11 @@ test_that("grouped_df errors on empty vars (#398)",{
   expect_error( m %>% do(mpg = mean(.$mpg)) )
 })
 
+test_that("group_by only creates one group for NA (#401)", {
+  x <- as.numeric(c(NA,NA,NA,10:1,10:1))
+  w <- c(20,30,40,1:10,1:10)*10
+  
+  n_distinct(x) # 11 OK
+  res <- data.frame(x=x,w=w) %>% group_by(x) %>% summarise(n=n())
+  expect_equal(nrow(res), 11L) 
+})